bless tests
