+++ tests/ui/async-await/issue-70818.rs
@@ -2,9 +2,8 @@
 
 use std::future::Future;
 fn foo<T: Send, U>(ty: T, ty1: U) -> impl Future<Output = (T, U)> + Send {
-    //~^ ERROR future cannot be sent between threads safely
+    //~^ ERROR: future cannot be sent between threads safely
     async { (ty, ty1) }
-    //~^ ERROR future cannot be sent between threads safely
 }
 
 fn main() {}
