+++ tests/ui/async-await/issue-70935-complex-spans.stderr
@@ -1,47 +1,4 @@
 error[E0277]: `*mut ()` cannot be shared between threads safely
-  --> $DIR/issue-70935-complex-spans.rs:17:5
-   |
-LL | /     async move {
-LL | |
-LL | |         baz(|| async {
-LL | |             foo(x.clone());
-LL | |         }).await;
-LL | |     }
-   | |_____^ `*mut ()` cannot be shared between threads safely
-   |
-   = help: within `NotSync`, the trait `Sync` is not implemented for `*mut ()`
-note: required because it appears within the type `PhantomData<*mut ()>`
-  --> $SRC_DIR/core/src/marker.rs:LL:COL
-note: required because it appears within the type `NotSync`
-  --> $DIR/issue-70935-complex-spans.rs:9:8
-   |
-LL | struct NotSync(PhantomData<*mut ()>);
-   |        ^^^^^^^
-   = note: required for `&NotSync` to implement `Send`
-note: required because it's used within this closure
-  --> $DIR/issue-70935-complex-spans.rs:19:13
-   |
-LL |         baz(|| async {
-   |             ^^
-note: required because it's used within this `async` fn body
-  --> $DIR/issue-70935-complex-spans.rs:12:67
-   |
-LL |   async fn baz<T>(_c: impl FnMut() -> T) where T: Future<Output=()> {
-   |  ___________________________________________________________________^
-LL | | }
-   | |_^
-note: required because it's used within this `async` block
-  --> $DIR/issue-70935-complex-spans.rs:17:5
-   |
-LL |     async move {
-   |     ^^^^^^^^^^
-note: required by a bound in an opaque type
-  --> $DIR/issue-70935-complex-spans.rs:15:37
-   |
-LL | fn foo(x: NotSync) -> impl Future + Send {
-   |                                     ^^^^
-
-error[E0277]: `*mut ()` cannot be shared between threads safely
   --> $DIR/issue-70935-complex-spans.rs:15:23
    |
 LL | fn foo(x: NotSync) -> impl Future + Send {
@@ -57,7 +14,7 @@ LL | struct NotSync(PhantomData<*mut ()>);
    |        ^^^^^^^
    = note: required for `&NotSync` to implement `Send`
 note: required because it's used within this closure
-  --> $DIR/issue-70935-complex-spans.rs:19:13
+  --> $DIR/issue-70935-complex-spans.rs:18:13
    |
 LL |         baz(|| async {
    |             ^^
@@ -74,6 +31,6 @@ note: required because it's used within this `async` block
 LL |     async move {
    |     ^^^^^^^^^^
 
-error: aborting due to 2 previous errors
+error: aborting due to 1 previous error
 
 For more information about this error, try `rustc --explain E0277`.
