+++ tests/ui/coroutine/issue-105084.stderr
@@ -1,27 +1,3 @@
-error[E0382]: borrow of moved value: `g`
-  --> $DIR/issue-105084.rs:38:14
-   |
-LL |     let mut g = #[coroutine]
-   |         ----- move occurs because `g` has type `{coroutine@$DIR/issue-105084.rs:16:5: 16:7}`, which does not implement the `Copy` trait
-...
-LL |     let mut h = copy(g);
-   |                      - value moved here
-...
-LL |     Pin::new(&mut g).resume(());
-   |              ^^^^^^ value borrowed here after move
-   |
-note: consider changing this parameter type in function `copy` to borrow instead if owning the value isn't necessary
-  --> $DIR/issue-105084.rs:10:21
-   |
-LL | fn copy<T: Copy>(x: T) -> T {
-   |    ----             ^ this parameter takes ownership of the value
-   |    |
-   |    in this function
-help: consider cloning the value if the performance cost is acceptable
-   |
-LL |     let mut h = copy(g.clone());
-   |                       ++++++++
-
 error[E0277]: the trait bound `Box<(i32, ())>: Copy` is not satisfied in `{coroutine@$DIR/issue-105084.rs:16:5: 16:7}`
   --> $DIR/issue-105084.rs:32:17
    |
@@ -45,7 +21,6 @@ note: required by a bound in `copy`
 LL | fn copy<T: Copy>(x: T) -> T {
    |            ^^^^ required by this bound in `copy`
 
-error: aborting due to 2 previous errors
+error: aborting due to 1 previous error
 
-Some errors have detailed explanations: E0277, E0382.
-For more information about an error, try `rustc --explain E0277`.
+For more information about this error, try `rustc --explain E0277`.
